organising files and struct renames to be consistent with spec prefix
- moved specResourceOperations to its own file
- moved specResponse to its own file
- moved specSchemaDefinition to its own file
- moved specSchemaDefinitionProperty to its own file
