+++ openapi/openapi_spec_resource_operations.go
@@ -1,0 +1,15 @@
+package openapi
+
+type specResourceOperations struct {
+	Post   *specResourceOperation
+	Get    *specResourceOperation
+	Put    *specResourceOperation
+	Delete *specResourceOperation
+}
+
+// specResourceOperation defines a resource operation
+type specResourceOperation struct {
+	SecuritySchemes  SpecSecuritySchemes
+	HeaderParameters SpecHeaderParameters
+	responses        specResponses
+}
